requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
pdf2image>=1.16.0
pymupdf>=1.23.0
accelerate>=0.20.0
